moved sound sending functions e.g. (PlayerQuitSoundController -> PlayerQuitMessageController)
